HARMONY-1736: Code refactoring.
