+++ services/harmony/app/frontends/ogc-coverages/index.ts
@@ -3,7 +3,6 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { Application, Response, Router } from 'express';
 import * as yaml from 'js-yaml';
-import { buildJsonErrorResponse, getCodeForError, getEndUserErrorMessage, getHttpStatusCode } from '../../util/errors';
 import getLandingPage from './get-landing-page';
 import getRequirementsClasses from './get-requirements-classes';
 
@@ -95,38 +94,3 @@ export function addOpenApiRoutes(app: Router): void {
   });
 }
 
-/**
- * Adds error handling appropriate to the OGC API to the given app
- * @param app - The express application which needs error handling routes
- */
-export function handleOpenApiErrors(app: Application): void {
-  app.use((err, req, res, next) => {
-    if (req.path.indexOf('/ogc-api-coverages/') === -1) {
-      next(err);
-      return;
-    }
-
-    let statusCode;
-    let message;
-    let code;
-    if (err.status && err.errors) {
-      // OpenAPI Validation errors;
-      statusCode = +err.status;
-      code = 'openapi.ValidationError';
-      const messages = err.errors.map((error) => `${error.location} parameter "${error.path}" ${error.message}`);
-      message = messages.join('\n\t');
-    } else {
-      statusCode = getHttpStatusCode(err);
-      code = getCodeForError(err);
-      message = getEndUserErrorMessage(err);
-    }
-    res.status(statusCode).json(buildJsonErrorResponse(code, message));
-
-    if (statusCode < 500) {
-      req.context.logger.error(`[${code}] ${message}`);
-    } else {
-      // Make sure we get stack traces when we throw an unexpected exception
-      req.context.logger.error(err);
-    }
-  });
-}
