+++ services/harmony/app/server.ts
@@ -13,9 +13,8 @@ import logForRoutes from './middleware/log-for-routes';
 import router, { RouterConfig } from './routers/router';
 import RequestContext from './models/request-context';
 import HarmonyRequest from './models/harmony-request';
-import * as ogcCoveragesApi from './frontends/ogc-coverages';
-import * as ogcEdrApi from './frontends/ogc-edr';
 import serviceResponseRouter from './routers/backend-router';
+import { handleOpenApiErrors } from './util/errors';
 import logger from './util/log';
 import * as exampleBackend from '../example/http-backend';
 import cmrCollectionReader from './middleware/cmr-collection-reader';
@@ -158,8 +157,7 @@ function buildFrontendServer(port: number, hostBinding: string, config: RouterCo
   app.use('/', router(config));
   // Error handlers that format errors outside of their routes / middleware need to be mounted
   // at the top level, not on a child router, or they get skipped.
-  ogcCoveragesApi.handleOpenApiErrors(app);
-  ogcEdrApi.handleOpenApiErrors(app);
+  handleOpenApiErrors(app);
   app.use(errorHandler);
 
   let listener;
